requirements.txt
Django==4.2.9
huesdk==1.7
aiohttp==3.8.5
cryptography==41.0.2
daphne==4.0.0
dj-stripe==2.8.3
Django==4.2.9
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-cors-headers==4.2.0
django-crispy-forms==2.0
django-filter==23.2
djangorestframework-simplejwt==5.3.0
django-versatileimagefield==3.0
django-formtools==2.4.1
django-ranged-response==0.2.0
django-recaptcha==4.0.0
django-storages==1.13.2
django-timezone-field==5.0
djangorestframework==3.14.0
mssql-django==1.3
sqlparse==0.4.3
python-dotenv==1.0.0
python-decouple==3.8
ipython==8.20.0